fix: regenerate query filters with correct types from field definitions
journal_key filter is now Option<String> (was Option<i64>), author_key
raw filter uses explicit type: string. Adds try-lock to geiger allow.

+++ src/adapters/db/queries/keyword.rs
@@ -13,7 +13,7 @@ use serde::Deserialize;
 #[derive(Filter, Debug, Default, Deserialize)]
 pub struct KeywordQuery {
     #[query(eq = "level")]
-    pub level: Option<i64>,
+    pub level: Option<i16>,
     #[query(like = "name")]
     pub name: Option<String>,
 }
